Final update to this branch. Feature is complete with unit test to back it up.

+++ Server/Controllers/SurveyController.cs
@@ -3,6 +3,7 @@ using System;
 using Microsoft.AspNetCore.Mvc;
 using MySql.Data.MySqlClient;
 using System.Collections.Generic;
+using System.Globalization;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -20,7 +21,7 @@ namespace schedulesUnitedHosted.Server.Controllers
         // Helper method, to get all responses to the given survey
         public List<Response> getAllResponses(int eventID)
         {
-            List<Response> ret = null;
+            List<Response> ret = new List<Response>();
             string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
 
             DBCon conGen = new DBCon(conString);
@@ -38,7 +39,7 @@ namespace schedulesUnitedHosted.Server.Controllers
                         DateTime availableDay = DateTime.Parse(reader["availableDay"].ToString());
                         int availableHour = Int32.Parse(reader["availableHour"].ToString());
 
-                        ret[i] = new Response(accountID, eventID, availableDay, availableHour);
+                        ret.Add(new Response(accountID, eventID, availableDay, availableHour));
                         i++;
                     }
                 con.Close();
@@ -50,7 +51,7 @@ namespace schedulesUnitedHosted.Server.Controllers
         [HttpGet("/surveys/{userID:int}")]
         public List<Survey> getAllSurvey(int userID)
         {
-            List<Survey> ret = null;
+            List<Survey> ret = new List<Survey>();
             string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
 
             DBCon conGen = new DBCon(conString);
@@ -71,7 +72,7 @@ namespace schedulesUnitedHosted.Server.Controllers
                         int host = Int32.Parse(reader["eventHost"].ToString());
                         List<Response> responses = getAllResponses(Int32.Parse(reader["eventID"].ToString()));
 
-                        ret[i] = new Survey(id, name, start, end, host, responses);
+                        ret.Add(new Survey(id, name, start, end, host, responses));
                         i++;
                     }
                 con.Close();
@@ -125,7 +126,7 @@ namespace schedulesUnitedHosted.Server.Controllers
             using (con)
             {
                 con.Open();
-                MySqlCommand getSurvey = new MySqlCommand($"SELECT eventID FROM CalendarEvents WHERE eventName = '{cleaned}' & eventHost = {ownerID}", con);
+                MySqlCommand getSurvey = new MySqlCommand($"SELECT eventID FROM CalendarEvents WHERE eventName = '{cleaned}' AND eventHost = {ownerID}", con);
                 using (var reader = getSurvey.ExecuteReader())
                     while (reader.Read())
                     {
@@ -140,15 +141,27 @@ namespace schedulesUnitedHosted.Server.Controllers
         [HttpGet("{surveyID:int}/{accountID:int}")]
         public List<Response> getUsersResponses(int surveyID, int accountID)
         {
-            List<Response> ret = getAllResponses(surveyID);
-            // iterate through all responses
-            for(int i = 0; i < ret.Count; i++)
+            List<Response> ret = new List<Response>();
+            string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
+
+            DBCon conGen = new DBCon(conString);
+            MySqlConnection con = conGen.GetConnection();
+
+            using (con)
             {
-                // if the response isn't from the requested user it removes it from the list
-                if(ret[i].AccId == accountID)
-                {
-                    ret.RemoveAt(i);
-                }
+                con.Open();
+                MySqlCommand getResponses = new MySqlCommand($"SELECT * FROM Availabilities WHERE eventID = {surveyID} AND accountID = {accountID}", con);
+                int i = 0;
+                using (var reader = getResponses.ExecuteReader())
+                    while (reader.Read())
+                    {
+                        DateTime availableDay = DateTime.Parse(reader["availableDay"].ToString());
+                        int availableHour = Int32.Parse(reader["availableHour"].ToString());
+
+                        ret.Add(new Response(accountID, surveyID, availableDay, availableHour));
+                        i++;
+                    }
+                con.Close();
             }
             return ret;
         }
@@ -164,8 +177,27 @@ namespace schedulesUnitedHosted.Server.Controllers
             {
                 con.Open();
                 //Create survey
-                MySqlCommand createSurvey = new MySqlCommand($"INSERT INTO Availabilities VALUES ({create.AccId}, {create.EventId}, '{create.Availability.ToString()}', {create.Hour})", con);
+                //throw new Exception($"INSERT INTO Availabilities VALUES ({create.AccId}, {create.EventId}, '{create.Availability.ToString("yyyy-MM-dd")}', {create.Hour})");
+                MySqlCommand createSurvey = new MySqlCommand($"INSERT INTO Availabilities VALUES ({create.AccId}, {create.EventId}, '{create.Availability.ToString("yyyy-MM-dd")}', {create.Hour})", con);
                 createSurvey.ExecuteNonQuery();
+                con.Close();
+            }
+        }
+
+        // POST <SurveyController>
+        [HttpPost("delete")]
+        public void deleteResponse([FromBody] Response delete)
+        {
+            string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
+            DBCon conGen = new DBCon(conString);
+            MySqlConnection con = conGen.GetConnection();
+            using (con)
+            {
+                con.Open();
+                //Create survey
+                //throw new Exception($"DELETE FROM Availabilities WHERE accountID = {delete.AccId} AND eventID = {delete.EventId} AND availableDay = '{delete.Availability.ToString("yyyy-MM-dd")}' AND availableHour = {delete.Hour}");
+                MySqlCommand deleteSurvey = new MySqlCommand($"DELETE FROM Availabilities WHERE accountID = {delete.AccId} AND eventID = {delete.EventId} AND availableDay = '{delete.Availability.ToString("yyyy-MM-dd")}' AND availableHour = {delete.Hour}", con);
+                deleteSurvey.ExecuteNonQuery();
                 con.Close();
             }
         }
@@ -181,7 +213,7 @@ namespace schedulesUnitedHosted.Server.Controllers
             {
                 con.Open();
                 //Create survey
-                MySqlCommand createSurvey = new MySqlCommand($"INSERT INTO CalendarEvents VALUES (NULL, '{create.name}', '{create.start}', '{create.end}', '{create.host}')", con);
+                MySqlCommand createSurvey = new MySqlCommand($"INSERT INTO CalendarEvents VALUES (NULL, '{create.name}', '{create.start.ToString("yyyy-MM-dd")}', '{create.end.ToString("yyyy-MM-dd")}', {create.host})", con);
                 createSurvey.ExecuteNonQuery();
                 con.Close();
             }
@@ -193,16 +225,8 @@ namespace schedulesUnitedHosted.Server.Controllers
         {
             User owner = combined.user;
             Survey edit = combined.survey;
-            DeleteSurvey(edit.id, owner);
-            CreateSurvey(edit);
-        }
-
-        // POST <SurveyController>
-        [HttpPost("delete/{id:int}")]
-        public void DeleteSurvey(int id, [FromBody] User owner)
-        {
             User cleaned = DBCon.Clean(owner);
-            if(getOneSurvey(id).host == cleaned.accountID)
+            if (getOneSurvey(edit.id).host == cleaned.accountID)
             {
                 string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
                 DBCon conGen = new DBCon(conString);
@@ -211,10 +235,45 @@ namespace schedulesUnitedHosted.Server.Controllers
                 {
                     con.Open();
                     //Delete survey
-                    MySqlCommand deleteSurvey = new MySqlCommand($"DELETE FROM CalendarEvents WHERE eventID = {id}", con);
-                    deleteSurvey.ExecuteNonQuery();
+                    MySqlCommand editSurvey = new MySqlCommand($"UPDATE CalendarEvents SET eventName = '{DBCon.Clean(edit.name)}', startDate = '{edit.start.ToString("yyyy-MM-dd")}', endDate = '{edit.end.ToString("yyyy-MM-dd")}' WHERE eventID = {edit.id}", con);
+                    editSurvey.ExecuteNonQuery();
                     con.Close();
                 }
+            }
+        }
+
+        // POST <SurveyController>
+        [HttpPost("delete/{id:int}")]
+        public void DeleteSurvey(int id, [FromBody] User owner)
+        {
+            User cleaned = DBCon.Clean(owner);
+            var survey = getOneSurvey(id);
+            if(survey != null)
+            {
+                if(survey.host == cleaned.accountID)
+                {
+                    string conString = Environment.GetEnvironmentVariable("DB_CONNECTION_STRING");
+                    DBCon conGen = new DBCon(conString);
+                    MySqlConnection con = conGen.GetConnection();
+                    using (con)
+                    {
+                        con.Open();
+                        //Delete survey
+                        MySqlCommand deleteSurvey = new MySqlCommand($"DELETE FROM CalendarEvents WHERE eventID = {id}", con);
+                        deleteSurvey.ExecuteNonQuery();
+                        deleteSurvey = new MySqlCommand($"DELETE FROM Availabilities WHERE eventID = {id}", con);
+                        deleteSurvey.ExecuteNonQuery();
+                        con.Close();
+                    }
+                }
+                else
+                {
+                    throw new Exception("User is not owner");
+                }
+            }
+            else
+            {
+                throw new Exception("Survey doesn't exist");
             }
         }
     }
